test: run empty file-list transfer control on all platforms
The empty-source-directory transfer control lived in the Unix-only
error_recovery_integration module (gated at the mod declaration), so it
silently skipped Windows even though its behaviour - a recursive transfer
of an empty tree exiting 0 - is platform-agnostic and needs no POSIX-only
surface. Its claimed coverage did not match its actual coverage.

Move it into a dedicated, ungated empty_flist_transfer module so the
end-to-end empty file-list exit-0 contract is exercised on every platform,
alongside the existing wire-level empty-flist coverage. The sibling tests
that genuinely require Unix (unreadable directories via permission bits)
stay gated.
